rename generated managers
