+++ src/management/__generated/managers/anomaly-manager.ts
