+++ src/management/__generated/managers/attack-protection-manager.ts
