+++ src/management/__generated/managers/branding-manager.ts
