+++ src/management/__generated/managers/emails-manager.ts
