+++ src/management/__generated/managers/hooks-manager.ts
