+++ src/management/__generated/managers/index.ts
@@ -1,31 +1,31 @@
-export * from './ActionsManager';
-export * from './AnomalyManager';
-export * from './AttackProtectionManager';
-export * from './BlacklistsManager';
-export * from './BrandingManager';
-export * from './ClientGrantsManager';
-export * from './ClientsManager';
-export * from './ConnectionsManager';
-export * from './CustomDomainsManager';
-export * from './DeviceCredentialsManager';
-export * from './EmailTemplatesManager';
-export * from './EmailsManager';
-export * from './GrantsManager';
-export * from './GuardianManager';
-export * from './HooksManager';
-export * from './JobsManager';
-export * from './KeysManager';
-export * from './LogStreamsManager';
-export * from './LogsManager';
-export * from './OrganizationsManager';
-export * from './PromptsManager';
-export * from './ResourceServersManager';
-export * from './RolesManager';
-export * from './RulesManager';
-export * from './RulesConfigsManager';
-export * from './StatsManager';
-export * from './TenantsManager';
-export * from './TicketsManager';
-export * from './UserBlocksManager';
-export * from './UsersManager';
-export * from './UsersByEmailManager';
+export * from './actions-manager';
+export * from './anomaly-manager';
+export * from './attack-protection-manager';
+export * from './blacklists-manager';
+export * from './branding-manager';
+export * from './client-grants-manager';
+export * from './clients-manager';
+export * from './connections-manager';
+export * from './custom-domains-manager';
+export * from './device-credentials-manager';
+export * from './email-templates-manager';
+export * from './emails-manager';
+export * from './grants-manager';
+export * from './guardian-manager';
+export * from './hooks-manager';
+export * from './jobs-manager';
+export * from './keys-manager';
+export * from './log-streams-manager';
+export * from './logs-manager';
+export * from './organizations-manager';
+export * from './prompts-manager';
+export * from './resource-servers-manager';
+export * from './roles-manager';
+export * from './rules-manager';
+export * from './rules-configs-manager';
+export * from './stats-manager';
+export * from './tenants-manager';
+export * from './tickets-manager';
+export * from './user-blocks-manager';
+export * from './users-manager';
+export * from './users-by-email-manager';
