+++ src/management/__generated/managers/jobs-manager.ts
