+++ src/management/__generated/managers/logs-manager.ts
