+++ src/management/__generated/managers/prompts-manager.ts
