+++ src/management/__generated/managers/rules-configs-manager.ts
