+++ src/management/__generated/managers/stats-manager.ts
